LayoutTests/http/tests/inspector/resource-parameters-expected.txt

Tests that resources panel shows form data parameters.

  

{
    pageref : "about:blank"
    startedDateTime : <string>
    time : <number>
    request : {
        method : "POST"
        url : "http://localhost:8000/inspector/resources/post-target.cgi?queryParam1=queryValue1&queryParam2="
        headers : <object>
        headersSize : -1
        bodySize : -1
        queryString : {
            0 : {
                name : "queryParam1"
                value : "queryValue1"
            }
            1 : {
                name : "queryParam2"
                value : ""
            }
        }
        postData : {
            mimeType : "application/x-www-form-urlencoded"
            text : "formParam1=formValue1&formParam2="
            params : {
                0 : {
                    name : "formParam1"
                    value : "formValue1"
                }
                1 : {
                    name : "formParam2"
                    value : ""
                }
            }
        }
    }
    response : {
        status : 200
        statusText : "OK"
        headers : <object>
        content : {
            size : 14
            mimeType : "application/xml"
        }
        redirectURL : ""
        headersSize : -1
        bodySize : 14
    }
    timings : {
        blocked : -1
        dns : -1
        connect : -1
        send : -1
        wait : <number>
        receive : <number>
        ssl : -1
    }
}
